unify

+++ arch/x86/64/include/ArchThreads.h
@@ -146,7 +146,7 @@ public:
  * @param userspace_register
  *
  */
-  static void printThreadRegisters(Thread *thread, uint32 userspace_registers, bool verbose = true);
+  static void printThreadRegisters(Thread *thread, size_t userspace_registers, bool verbose = true);
   static void printThreadRegisters(Thread *thread, bool verbose = true);
 
   /**
